+++ src/simulation/architecture/system_model/sim_model.cpp
@@ -423,9 +423,9 @@ std::set<std::pair<long int, long int>> SimModel::getMessageExchangeData(std::st
 {
     std::set<std::pair<long int, long int>> returnPairs;
     bool messageFound = false;
-    for(uint64_t i=0; i<SystemMessaging::GetInstance()->getProcessCount(); i++)
+    for(int64_t i=0; i<SystemMessaging::GetInstance()->getProcessCount(); i++)
     {
-        if(procList.find(i) == procList.end() && procList.size() > 0)
+        if(procList.find((uint64_t)i) == procList.end() && procList.size() > 0)
         {
             continue;
         }
